% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jive_est.R
\name{jive_est}
\alias{jive_est}
\title{JIVE Estimator}
\usage{
jive_est(y, X, Z, type = UJIVE1, vcov = HCE)
}
\arguments{
\item{y:}{a n x 1 matrix of the dependent variable,}

\item{X:}{a n x k matrix of independent variables,}

\item{Z:}{a n x l matrix of instrumental variables with l >= k,}

\item{type:}{UJIVE1 or UJIVE2 for the type of estimator that should be used,}

\item{vcov:}{HC0 for homoskedastic or HCE for heteroskedastic robust standard errors.}
}
\value{
A list of regression results
}
\description{
This function calculates estimands by using the JIVE estimator
as proposed by Angrist, Imbens, Krueger 1999. The output includes the coefficents,
variance matrix, standard errors and test statistics.
The type of estimator (i.e. UJIVE1 vs. UJIVE2) and standard errors
(i.e. homoskedastic: HC0 and heteroskedastic robust: HCE) can be given as parameters.
}
\details{
Version: 0.1
}
